graphics: Optimize vertices function for browsers
