+++ imageparts.go
@@ -17,8 +17,6 @@ package ebiten
 import (
 	"image"
 	"math"
-
-	"github.com/hajimehoshi/ebiten/internal/graphics"
 )
 
 // An ImagePart is deprecated (as of 1.1.0-alpha): Use ImageParts instead.
@@ -74,64 +72,4 @@ func u(x, width2p int) int16 {
 
 func v(y, height2p int) int16 {
 	return int16(math.MaxInt16 * y / height2p)
-}
-
-func vertices(parts ImageParts, width, height int, geo *GeoM) []int16 {
-	// TODO: This function should be in graphics package?
-	totalSize := graphics.QuadVertexSizeInBytes() / 2
-	oneSize := totalSize / 4
-	l := parts.Len()
-	vs := make([]int16, l*totalSize)
-	width2p := graphics.NextPowerOf2Int(width)
-	height2p := graphics.NextPowerOf2Int(height)
-	n := 0
-	geo16 := floatsToInt16s(geo.Element(0, 0),
-		geo.Element(0, 1),
-		geo.Element(1, 0),
-		geo.Element(1, 1),
-		geo.Element(0, 2),
-		geo.Element(1, 2))
-	for i := 0; i < l; i++ {
-		dx0, dy0, dx1, dy1 := parts.Dst(i)
-		if dx0 == dx1 || dy0 == dy1 {
-			continue
-		}
-		x0, y0, x1, y1 := int16(dx0), int16(dy0), int16(dx1), int16(dy1)
-		sx0, sy0, sx1, sy1 := parts.Src(i)
-		if sx0 == sx1 || sy0 == sy1 {
-			continue
-		}
-		u0, v0, u1, v1 := u(sx0, width2p), v(sy0, height2p), u(sx1, width2p), v(sy1, height2p)
-		offset := n * totalSize
-		vs[offset] = x0
-		vs[offset+1] = y0
-		vs[offset+2] = u0
-		vs[offset+3] = v0
-		for j, g := range geo16 {
-			vs[offset+4+j] = g
-		}
-		vs[offset+oneSize] = x1
-		vs[offset+oneSize+1] = y0
-		vs[offset+oneSize+2] = u1
-		vs[offset+oneSize+3] = v0
-		for j, g := range geo16 {
-			vs[offset+oneSize+4+j] = g
-		}
-		vs[offset+2*oneSize] = x0
-		vs[offset+2*oneSize+1] = y1
-		vs[offset+2*oneSize+2] = u0
-		vs[offset+2*oneSize+3] = v1
-		for j, g := range geo16 {
-			vs[offset+2*oneSize+4+j] = g
-		}
-		vs[offset+3*oneSize] = x1
-		vs[offset+3*oneSize+1] = y1
-		vs[offset+3*oneSize+2] = u1
-		vs[offset+3*oneSize+3] = v1
-		for j, g := range geo16 {
-			vs[offset+3*oneSize+4+j] = g
-		}
-		n++
-	}
-	return vs[:n*totalSize]
 }
