--- math_js.go
@@ -1,31 +1,0 @@
-// Copyright 2016 The Ebiten Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// +build js
-
-package ebiten
-
-import (
-	"github.com/gopherjs/gopherjs/js"
-)
-
-func floatsToInt16s(xs ...float64) []int16 {
-	a := js.Global.Get("ArrayBuffer").New(4 * len(xs))
-	af32 := js.Global.Get("Float32Array").New(a)
-	a16 := js.Global.Get("Int16Array").New(a)
-	for i, x := range xs {
-		af32.SetIndex(i, x)
-	}
-	return a16.Interface().([]int16)
-}
